restraapp/backend/requirements.txt
fastapi
uvicorn
pydantic[email]
pydantic-settings
python-dotenv
pymongo

passlib[bcrypt] 
python-jose
bcrypt==4.3.0